requirements.txt
beautifulsoup4==4.12.2
faiss-cpu==1.10.0
networkx==3.1
nltk==3.6.7
numpy==1.26.4
pandas==1.4.4
scikit-learn==1.3.0
tqdm==4.67.1
torch==1.12.1
torch-geometric==2.1.0.post1